(SERVER-1469) Add dynamic timestamp and transaction-uuid
This commit adds code to the `process_gatling_recording` script
that allows it to handle the step about dynamically injecting
a timestamp and transaction-uuid into the report.

It was easier to implement this if this step happens *after* the
step about renaming HTTP requests, so I reordered a few of the
steps in order to implement this.